requirements.txt
python-dotenv
pydantic
requests
langchain
langchain-google-genai
